Make Abstract{}Builder inner classes Abstract{}.Builder
