+++ sdk/src/test/java/io/opentelemetry/sdk/metrics/AbstractCounterBuilderTest.java
@@ -30,7 +30,7 @@ import org.junit.rules.ExpectedException;
 import org.junit.runner.RunWith;
 import org.junit.runners.JUnit4;
 
-/** Unit tests for {@link AbstractCounterBuilder}. */
+/** Unit tests for {@link AbstractCounter.Builder}. */
 @RunWith(JUnit4.class)
 public class AbstractCounterBuilderTest {
   @Rule public ExpectedException thrown = ExpectedException.none();
@@ -68,7 +68,7 @@ public class AbstractCounterBuilderTest {
   }
 
   private static final class TestInstrumentBuilder
-      extends AbstractCounterBuilder<TestInstrumentBuilder, TestInstrument> {
+      extends AbstractCounter.Builder<TestInstrumentBuilder, TestInstrument> {
     TestInstrumentBuilder(
         String name,
         MeterSharedState sharedState,
