+++ sdk/src/test/java/io/opentelemetry/sdk/metrics/AbstractMeasureBuilderTest.java
@@ -30,7 +30,7 @@ import org.junit.rules.ExpectedException;
 import org.junit.runner.RunWith;
 import org.junit.runners.JUnit4;
 
-/** Unit tests for {@link AbstractMeasureBuilder}. */
+/** Unit tests for {@link AbstractMeasure.Builder}. */
 @RunWith(JUnit4.class)
 public class AbstractMeasureBuilderTest {
   @Rule public ExpectedException thrown = ExpectedException.none();
@@ -68,7 +68,7 @@ public class AbstractMeasureBuilderTest {
   }
 
   private static final class TestInstrumentBuilder
-      extends AbstractMeasureBuilder<TestInstrumentBuilder, TestInstrument> {
+      extends AbstractMeasure.Builder<TestInstrumentBuilder, TestInstrument> {
     TestInstrumentBuilder(
         String name,
         MeterSharedState sharedState,
